- Adicionando Repositories - Utilizando tabela de dados pré consolidados para exibir gráficos - Inserir ou remover movimentações atualiza a tabela de dados consolidados - Adicionando seleção de ano nos gráficos - Gráfico de Custos por categoria agora só exibe um nível por vez - Adicionado opção de exibir gráficos por categoria - Adicionado opção de exibir dados que compoem o gráfico da categoria

+++ app/database/migrations/2020_08_13_033812_create_entries_table.php
@@ -14,7 +14,7 @@ class CreateEntriesTable extends Migration {
             $table->foreignId('group_id')
                 ->constrained();
 
-            $table->float('amount');
+            $table->double('amount');
             $table->date('at');
 
             $table->string('description')
